start and endate working + tickets semi working

+++ 2NMCT3_Waeytens_Wouter_BAP_01/eindopdracht/viewmodel/SettingsTicketsVM.cs
@@ -7,6 +7,7 @@ using System.Collections.ObjectModel;
 using eindopdracht.model;
 using System.Windows.Input;
 using GalaSoft.MvvmLight.Command;
+using System.Windows;
 
 namespace eindopdracht.viewmodel
 {
@@ -40,8 +41,24 @@ namespace eindopdracht.viewmodel
         {
             get { return _SaveTicketType; }
             set { _SaveTicketType = value; OnPropertyChanged("SaveTicketType"); }
-        }      
+        }
 
+        private Ticket _reserveTicket;
+
+        public Ticket ReserveTicket
+        {
+            get { return _reserveTicket; }
+            set { _reserveTicket = value; OnPropertyChanged("ReserveTicket"); }
+        }
+
+        private ObservableCollection<int> _lstAantalTickets;
+
+        public ObservableCollection<int> lstAantalTickets
+        {
+            get { return _lstAantalTickets; }
+            set { _lstAantalTickets = value; OnPropertyChanged("lstAantalTickets"); }
+        }
+        
         private String _textEditSaveType;
 
         public String TextEditSaveType
@@ -66,10 +83,22 @@ namespace eindopdracht.viewmodel
             _TicketList = Ticket.GetTickets();
             _ticketTypeList = TicketType.GetTypesTickets();
             SaveTicketType = new TicketType();
+            ReserveTicket = new Ticket();
             TextEditSaveType = "Save";
             visibleButton = "Hidden";
+            lstAantalTickets = aantalTicketsFill();
         }
         #endregion
+
+        private ObservableCollection<int> aantalTicketsFill()
+        {
+            ObservableCollection<int> lst = new ObservableCollection<int>();
+            for (int i = 1; i < 11; i++)
+            {
+                lst.Add(i);
+            }
+            return lst;
+        }
 
         #region commands
         public ICommand addTicketTypeCommand
@@ -88,7 +117,6 @@ namespace eindopdracht.viewmodel
                 TicketType.EditTicketType(SaveTicketType);
             }
             TicketTypeList = TicketType.GetTypesTickets();
-
         }
 
         public ICommand editTypeCommand
@@ -128,8 +156,32 @@ namespace eindopdracht.viewmodel
             SaveTicketType = new TicketType();
             visibleButton = "Hidden";
             TextEditSaveType = "Save";
-            TicketType.DeleteTicketType(type);
+            //alle gereserveerde tickets met dat type ook verwijderen en waarschuwen met een dialogbox
+            string sMessageBoxText = "Weet je zeker dat je het Type "+type.Name+" wil verwijderen alle gereserveerde tickets voor het type "+type.Name+" worden ook verwijderd.";
+            string sCaption = "Bent u zeker?";
+
+            MessageBoxButton btnMessageBox = MessageBoxButton.YesNo;
+            MessageBoxImage icnMessageBox = MessageBoxImage.Asterisk;
+
+            MessageBoxResult rsltMessageBox = MessageBox.Show(sMessageBoxText, sCaption, btnMessageBox, icnMessageBox);
+            if (rsltMessageBox == MessageBoxResult.Yes)
+            {
+                TicketType.DeleteTicketType(type);
+            }
+
             TicketTypeList = TicketType.GetTypesTickets();
+        }
+
+        public ICommand ReserveTicketCommand
+        {
+            get { return new RelayCommand(reserveHandler); }
+        }
+
+        private void reserveHandler()
+        {
+            Ticket.ReserveTicket(ReserveTicket);
+            MessageBox.Show("Ticket Reserved for " + ReserveTicket.TicketHolder);
+            ReserveTicket = new Ticket();
         }
 
         #endregion
